Cleaned up connection manager (moved many classes to their own files).
